requirements.txt
astroid==2.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
entrypoints==0.3
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.0
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.15.4
pandas==0.24.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.3.1
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
pyzmq==17.1.2
requests==2.21.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.6.2
SQLAlchemy==1.2.15
-e git+https://github.com/billwarker/feature-film-analysis.git@e1e39863ce32d158b4614e64069c60a67f3a5aec#egg=src
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
